Added Additional Functions. !Zscore plot is not working!)
Added Plot behavior and zscore plot to the Gui.

Both the original functions in FiberClass.py were updated significantly. Behaviors are now imported into columns with strings 'S' signifies the start of a point or state behavior, 'O' means that a state behavior is occuring and 'E' marks the end of a state behavior.

Plot Zscore will now work with state or point behaviors. It is not running currently due to bugs.