Adding Try Catches to the notebook
Haven't updated the script yet. Mostly rearranging the functions so they catch inside the loop instead of outside it.

Haven't updated try catches for behavior, I think I'm going to do some work on the behavior reading first.

Need to add more specific try catches to some of the functions in the future